requirements.txt
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
dnspython==2.7.0
idna==3.10
numpy==2.0.2
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.3
six==1.17.0
soupsieve==2.7
tabulate==0.9.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
